% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_EBA_page.R
\name{parse_EBA_page}
\alias{parse_EBA_page}
\title{Parse EBA rulebook page}
\usage{
parse_EBA_page()
}
\value{
Data frame with Title/Scope, URL, and Rule ID (extracted from URLs)
}
\description{
Parse EBA rulebook page
}
\examples{
\dontrun{
parse_EBA_page()
}
}
