% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_EBA_QA_page.R
\name{parse_EBA_QA_page}
\alias{parse_EBA_QA_page}
\title{Parse EBA Q&A pages}
\usage{
parse_EBA_QA_page()
}
\value{
Data frame with Q&A Title (Subject Matter) and corresponding URL
}
\description{
Parse EBA Q&A pages
}
\examples{
\dontrun{
qa_total <- parse_EBA_QA_page()
saveRDS(qa_total, here::here("data", "EBA_QA.Rds"))
}
}
