% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_EBA.R
\name{scrape_EBA}
\alias{scrape_EBA}
\title{Get the content of the EBA single rulebook}
\usage{
scrape_EBA(x)
}
\arguments{
\item{x}{String. Optional URL of rules to scrape.}
}
\value{
Data frame with the content of EBA rulebook.
Additional columns (QA to DA_QA) show the count of occurrences of tags.
}
\description{
The function first checks if scraped data is available locally.
If so, the rulebook is not scraped.
}
\examples{
\dontrun{
# when rulebook URLs are saved locally
scrape_EBA()
# if not, rules will be scraped when URL is provided
scrape_EBA("https://eba.europa.eu/regulation-and-policy/single-rulebook/interactive-single-rulebook/-/interactive-single-rulebook/toc/504/article-id/601")
}
}
